7 records

+++ src/cz/cuni/mff/d3s/trupple/language/parser/identifierstable/types/constant/RealConstantDescriptor.java
@@ -12,7 +12,7 @@ public class RealConstantDescriptor implements ConstantDescriptor {
 
     @Override
     public FrameSlotKind getSlotKind() {
-        return FrameSlotKind.Long;
+        return FrameSlotKind.Double;
     }
 
     @Override
